Shayla and DaShawn
Completion of the Project:
We worked together on each step. For the most part DaShawn did the code and Shayla created the layout. Shayla included pictures on the main page making the site appealing. We planned the layout in the past, so he coded it based on the drafts. The site's format has indeed changed from that of what we first started off with; however, we tried our best to keep some of the main components and major layouts of the website. We used html and Css on GitHub (Sandbox). Some other sources that helped us along the way were w3schools. It gave us code elements to get ideas on how the code should be written.